+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Added
 - Feed me support for dateCreated and dateUpdated
+- Dashboard widget
 
 ## 3.0.0.1 - 2020-01-28
 ### Fixed